refactor: 패키지 구조 및 위치 변경
