+++ src/main/java/woowacourse/chatting/ChattingApplication.java
@@ -2,16 +2,10 @@ package woowacourse.chatting;
 
 import jakarta.annotation.PostConstruct;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import woowacourse.chatting.domain.message.ChatRoom;
 import woowacourse.chatting.dto.AddMemberRequest;
-import woowacourse.chatting.repository.message.ChatRoomRepository;
 import woowacourse.chatting.service.MemberService;
-
-import java.util.List;
-import java.util.UUID;
 
 @SpringBootApplication
 public class ChattingApplication {
