+++ src/main/java/woowacourse/chatting/controller/AuthController.java
@@ -11,8 +11,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import woowacourse.chatting.dto.AddMemberRequest;
 import woowacourse.chatting.dto.ResponseDto;
-import woowacourse.chatting.dto.ResponseToken;
-import woowacourse.chatting.dto.SignInDto;
+import woowacourse.chatting.dto.auth.ResponseToken;
+import woowacourse.chatting.dto.auth.SignInDto;
 import woowacourse.chatting.jwt.JwtToken;
 import woowacourse.chatting.service.AuthService;
 import woowacourse.chatting.service.MemberServiceImp;
