+++ src/main/java/woowacourse/chatting/controller/JwtController.java
@@ -7,7 +7,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
-import woowacourse.chatting.dto.ResponseToken;
+import woowacourse.chatting.dto.auth.ResponseToken;
 import woowacourse.chatting.jwt.JwtTokenProvider;
 import woowacourse.chatting.service.RefreshTokeService;
