+++ src/main/java/woowacourse/chatting/controller/chat/ChatController.java
@@ -10,13 +10,13 @@ import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import woowacourse.chatting.ChatRoomCache;
 import woowacourse.chatting.domain.Member;
-import woowacourse.chatting.domain.message.ChatMessage;
-import woowacourse.chatting.domain.message.ChatRoom;
-import woowacourse.chatting.domain.message.MessageType;
-import woowacourse.chatting.dto.message.ChatMessageDto;
-import woowacourse.chatting.dto.message.PrivateMessageDto;
-import woowacourse.chatting.repository.message.ChatMessageRepository;
-import woowacourse.chatting.repository.message.ChatRoomRepository;
+import woowacourse.chatting.domain.chat.ChatMessage;
+import woowacourse.chatting.domain.chat.ChatRoom;
+import woowacourse.chatting.domain.chat.MessageType;
+import woowacourse.chatting.dto.chat.ChatMessageDto;
+import woowacourse.chatting.dto.chat.PrivateMessageDto;
+import woowacourse.chatting.repository.chat.ChatMessageRepository;
+import woowacourse.chatting.repository.chat.ChatRoomRepository;
 import woowacourse.chatting.service.webSocket.ConnectedUserService;
 
 import java.security.Principal;
