+++ src/main/java/woowacourse/chatting/domain/chat/MessageType.java
@@ -1,4 +1,4 @@
-package woowacourse.chatting.domain.message;
+package woowacourse.chatting.domain.chat;
 
 public enum MessageType {
     PRIVATE,
