+++ src/main/java/woowacourse/chatting/repository/RefreshTokeRepository.java
@@ -3,7 +3,7 @@ package woowacourse.chatting.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
-import woowacourse.chatting.domain.RefreshToken;
+import woowacourse.chatting.domain.auth.RefreshToken;
 
 import java.util.Optional;
